+++ test-common/src/test/java/io/github/chrimle/o2jrm/tests/IGeneratedEnumTests.java
@@ -1,5 +1,6 @@
 package io.github.chrimle.o2jrm.tests;
 
+import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonValue;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonParser;
@@ -7,11 +8,17 @@ import com.google.gson.annotations.JsonAdapter;
 import com.google.gson.stream.JsonReader;
 import com.google.gson.stream.JsonWriter;
 import io.github.chrimle.o2jrm.GeneratedSource;
+import io.github.chrimle.o2jrm.annotations.TestAnnotationOne;
+import io.github.chrimle.o2jrm.annotations.TestAnnotationThree;
+import io.github.chrimle.o2jrm.annotations.TestAnnotationTwo;
 import io.github.chrimle.o2jrm.models.GeneratedField;
-import io.github.chrimle.o2jrm.tests.IGeneratedEnumTests.IOpenAPITests.ISchemaTests;
-import io.github.chrimle.o2jrm.tests.IGeneratedEnumTests.IOpenAPITests.ISchemaTests.IDeprecatedTests;
-import io.github.chrimle.o2jrm.tests.IGeneratedEnumTests.IOpenAPITests.ISchemaTests.IEnumTests;
-import io.github.chrimle.o2jrm.tests.IGeneratedEnumTests.IOpenAPITests.ISchemaTests.IEnumTests.IConstantsTests;
+import io.github.chrimle.o2jrm.tests.IGeneratedEnumTests.GeneratorConfigurationTests.ConfigOptionsTests;
+import io.github.chrimle.o2jrm.tests.IGeneratedEnumTests.GeneratorConfigurationTests.ConfigOptionsTests.AdditionalEnumTypeAnnotationsTests;
+import io.github.chrimle.o2jrm.tests.IGeneratedEnumTests.GeneratorConfigurationTests.ConfigOptionsTests.UseEnumCaseInsensitiveTests;
+import io.github.chrimle.o2jrm.tests.IGeneratedEnumTests.OpenAPITests.SchemaTests;
+import io.github.chrimle.o2jrm.tests.IGeneratedEnumTests.OpenAPITests.SchemaTests.DeprecatedTests;
+import io.github.chrimle.o2jrm.tests.IGeneratedEnumTests.OpenAPITests.SchemaTests.EnumTests;
+import io.github.chrimle.o2jrm.tests.IGeneratedEnumTests.OpenAPITests.SchemaTests.EnumTests.ConstantsTests;
 import io.github.chrimle.o2jrm.utils.CustomAssertions;
 import java.io.OutputStream;
 import java.io.OutputStreamWriter;
@@ -40,18 +47,18 @@ import org.junit.jupiter.params.provider.MethodSource;
  *
  * <p><b>OpenAPI Specification</b>
  *
- * <p>{@link IOpenAPITests}
+ * <p>{@link OpenAPITests}
  *
  * <ul>
- *   <li>{@link ISchemaTests `components.schemas.{schema}`}
- *   <li>{@link IDeprecatedTests `components.schemas.{schema}.deprecated`}
- *   <li>{@link IEnumTests `components.schemas.{schema}.enum`}
- *   <li>{@link IConstantsTests `components.schemas.{schema}.enum.{constants}`}
+ *   <li>{@link SchemaTests `components.schemas.{schema}`}
+ *   <li>{@link DeprecatedTests `components.schemas.{schema}.deprecated`}
+ *   <li>{@link EnumTests `components.schemas.{schema}.enum`}
+ *   <li>{@link ConstantsTests `components.schemas.{schema}.enum.{constants}`}
  * </ul>
  *
  * <p><b>`openapi-generator` Configurations</b>
  *
- * <p>{@link IGeneratorConfigurationTests}
+ * <p>{@link GeneratorConfigurationTests}
  *
  * <ul>
  *   <li>{@link ConfigOptionsTests configOptions}
@@ -67,15 +74,15 @@ public abstract class IGeneratedEnumTests {
 
   @Nested
   @DisplayName("Testing OpenAPI Schemas & Properties")
-  class IOpenAPITests {
+  class OpenAPITests {
 
     @Nested
     @DisplayName("Testing `components.schemas.{schema}`")
-    class ISchemaTests {
+    class SchemaTests {
 
       @Nested
       @DisplayName("Testing `components.schemas.{schema}.enum`")
-      class IEnumTests {
+      class EnumTests {
 
         @ParameterizedTest
         @MethodSource(GENERATED_ENUM_TESTS_METHOD_SOURCE)
@@ -86,7 +93,7 @@ public abstract class IGeneratedEnumTests {
 
         @Nested
         @DisplayName("Testing `getValue()` method")
-        class IGetValueMethodTests {
+        class GetValueMethodTests {
 
           @ParameterizedTest
           @MethodSource(GENERATED_ENUM_TESTS_METHOD_SOURCE)
@@ -126,7 +133,7 @@ public abstract class IGeneratedEnumTests {
 
         @Nested
         @DisplayName("Testing `components.schemas.{schema}.enum.{constants}`")
-        class IConstantsTests {
+        class ConstantsTests {
 
           @ParameterizedTest
           @MethodSource(GENERATED_ENUM_TESTS_METHOD_SOURCE)
@@ -169,7 +176,7 @@ public abstract class IGeneratedEnumTests {
 
       @Nested
       @DisplayName("Testing `components.schemas.{schema}.deprecated`")
-      class IDeprecatedTests {
+      class DeprecatedTests {
 
         @Nested
         @DisplayName("Testing `components.schemas.{schema}.deprecated: false`")
@@ -204,15 +211,15 @@ public abstract class IGeneratedEnumTests {
 
   @Nested
   @DisplayName("Testing `openapi-generator` Configurations & ConfigOptions")
-  class IGeneratorConfigurationTests {
+  class GeneratorConfigurationTests {
 
     @Nested
     @DisplayName("Testing `<library>`")
-    class ILibraryTests {
+    class LibraryTests {
 
       @Nested
       @DisplayName("Testing `<library>okhttp-gson</library>` (default)")
-      class IOkHttpGsonTests {
+      class OkHttpGsonTests {
 
         @ParameterizedTest
         @MethodSource(GENERATED_ENUM_TESTS_METHOD_SOURCE)
@@ -284,7 +291,7 @@ public abstract class IGeneratedEnumTests {
 
         @Nested
         @DisplayName("Testing the `Adapter` inner-class")
-        class IAdapterInnerClassTests {
+        class AdapterInnerClassTests {
 
           @ParameterizedTest
           @MethodSource(GENERATED_ENUM_TESTS_METHOD_SOURCE)
@@ -315,7 +322,7 @@ public abstract class IGeneratedEnumTests {
 
           @Nested
           @DisplayName("Testing the `write(JsonWriter, T)`-method")
-          class IWriteMethodTests {
+          class WriteMethodTests {
 
             @ParameterizedTest
             @MethodSource(GENERATED_ENUM_TESTS_METHOD_SOURCE)
@@ -376,7 +383,7 @@ public abstract class IGeneratedEnumTests {
 
           @Nested
           @DisplayName("Testing the `read(JsonReader)`-method")
-          class IReadMethodTests {
+          class ReadMethodTests {
 
             @ParameterizedTest
             @MethodSource(GENERATED_ENUM_TESTS_METHOD_SOURCE)
@@ -459,7 +466,7 @@ public abstract class IGeneratedEnumTests {
 
       @Nested
       @DisplayName("Testing `<library>webclient</library>`")
-      class IWebClientTests {
+      class WebClientTests {
 
         @ParameterizedTest
         @MethodSource(GENERATED_ENUM_TESTS_METHOD_SOURCE)
@@ -492,6 +499,473 @@ public abstract class IGeneratedEnumTests {
 
           CustomAssertions.assertClassIsNotAnnotatedWith(
               generatedSource.getClassUnderTest(), JsonAdapter.class);
+        }
+      }
+    }
+
+    @Nested
+    @DisplayName("Testing `<configOptions>`")
+    class ConfigOptionsTests {
+
+      @Nested
+      @DisplayName("Testing `<additionalEnumTypeAnnotations>`")
+      class AdditionalEnumTypeAnnotationsTests {
+
+        @Nested
+        @DisplayName("Testing `<additionalEnumTypeAnnotations></additionalEnumTypeAnnotations>`")
+        class AdditionalEnumTypeAnnotationsUnsetTests {
+
+          @ParameterizedTest
+          @MethodSource(GENERATED_ENUM_TESTS_METHOD_SOURCE)
+          @DisplayName("Generated `enum` class is NOT annotated with additional annotations")
+          void
+              whenConfigOptionAdditionalEnumTypeAnnotationsIsUnsetThenGeneratedEnumDoesNotHaveAdditionalAnnotations(
+                  final GeneratedSource generatedSource) {
+            Assumptions.assumeFalse(generatedSource.hasAdditionalEnumTypeAnnotations());
+
+            CustomAssertions.assertClassIsNotAnnotatedWith(
+                generatedSource.getClassUnderTest(), TestAnnotationOne.class);
+            CustomAssertions.assertClassIsNotAnnotatedWith(
+                generatedSource.getClassUnderTest(), TestAnnotationTwo.class);
+            CustomAssertions.assertClassIsNotAnnotatedWith(
+                generatedSource.getClassUnderTest(), TestAnnotationThree.class);
+          }
+        }
+
+        @Nested
+        @DisplayName(
+            "Testing `<additionalEnumTypeAnnotations>@TestAnnotationOne;@TestAnnotationTwo;@TestAnnotationThree</additionalEnumTypeAnnotations>`")
+        class AdditionalEnumTypeAnnotationsSetTests {
+
+          @ParameterizedTest
+          @MethodSource(GENERATED_ENUM_TESTS_METHOD_SOURCE)
+          @DisplayName("Generated `enum` class is annotated with additional annotations")
+          void
+              whenConfigOptionAdditionalEnumTypeAnnotationsIsSetThenGeneratedEnumHasAdditionalAnnotations(
+                  final GeneratedSource generatedSource) {
+            Assumptions.assumeTrue(generatedSource.hasAdditionalEnumTypeAnnotations());
+
+            CustomAssertions.assertClassIsAnnotatedWith(
+                generatedSource.getClassUnderTest(), TestAnnotationOne.class);
+            CustomAssertions.assertClassIsAnnotatedWith(
+                generatedSource.getClassUnderTest(), TestAnnotationTwo.class);
+            CustomAssertions.assertClassIsAnnotatedWith(
+                generatedSource.getClassUnderTest(), TestAnnotationThree.class);
+          }
+        }
+      }
+
+      @Nested
+      @DisplayName("Testing `<useEnumCaseInsensitive>`")
+      class UseEnumCaseInsensitiveTests {
+
+        @ParameterizedTest
+        @MethodSource(GENERATED_ENUM_TESTS_METHOD_SOURCE)
+        @DisplayName("Generated `enum` class ALWAYS has a `static fromValue(T)` method")
+        void alwaysGenerateEnumClassWithStaticFromValueMethod(
+            final GeneratedSource generatedSource) {
+          CustomAssertions.assertClassHasMethod(
+              generatedSource.getClassUnderTest(), "fromValue", generatedSource.fieldClasses()[0]);
+        }
+
+        @ParameterizedTest
+        @MethodSource(GENERATED_ENUM_TESTS_METHOD_SOURCE)
+        @DisplayName(
+            "When `serializationLibrary` is `jackson` Then generated `fromValue(T)` method is annotated with `@JsonCreator`")
+        void whenSerializationLibraryIsJacksonThenFromValueMethodIsAnnotatedWithJsonCreator(
+            final GeneratedSource generatedSource) {
+          Assumptions.assumeTrue(generatedSource.isSerializationLibraryJackson());
+
+          CustomAssertions.assertMethodIsAnnotatedWith(
+              CustomAssertions.assertClassHasMethod(
+                  generatedSource.getClassUnderTest(),
+                  "fromValue",
+                  generatedSource.fieldClasses()[0]),
+              JsonCreator.class);
+        }
+
+        @ParameterizedTest
+        @MethodSource(GENERATED_ENUM_TESTS_METHOD_SOURCE)
+        @DisplayName(
+            "When `serializationLibrary` is NOT `jackson` Then generated `fromValue(T)` method is NOT annotated with `@JsonCreator`")
+        void whenSerializationLibraryIsNotJacksonThenFromValueMethodIsNotAnnotatedWithJsonCreator(
+            final GeneratedSource generatedSource) {
+          Assumptions.assumeFalse(generatedSource.isSerializationLibraryJackson());
+
+          CustomAssertions.assertMethodIsNotAnnotatedWith(
+              CustomAssertions.assertClassHasMethod(
+                  generatedSource.getClassUnderTest(),
+                  "fromValue",
+                  generatedSource.fieldClasses()[0]),
+              JsonCreator.class);
+        }
+
+        @ParameterizedTest
+        @MethodSource(GENERATED_ENUM_TESTS_METHOD_SOURCE)
+        @DisplayName(
+            "Generated `static fromValue(T)` method ALWAYS `throw IllegalArgumentException` when `null` is given")
+        void alwaysThrowIllegalArgumentExceptionWhenProvidingNullAsArgumentToStaticFromValueMethod(
+            final GeneratedSource generatedSource) {
+          Assumptions.assumeFalse(generatedSource.enumUnknownDefaultCase());
+          CustomAssertions.assertStaticMethodThrowsWhenInvoked(
+              CustomAssertions.assertClassHasMethod(
+                  generatedSource.getClassUnderTest(),
+                  "fromValue",
+                  generatedSource.fieldClasses()[0]),
+              IllegalArgumentException.class,
+              (Object) null);
+        }
+
+        @ParameterizedTest
+        @MethodSource(GENERATED_ENUM_TESTS_METHOD_SOURCE)
+        @DisplayName(
+            "Generated `static fromValue(T)` method ALWAYS `throw IllegalArgumentException` when non-matching string `value` is given")
+        void
+            alwaysThrowIllegalArgumentExceptionWhenProvidingNonMatchingStringValueAsArgumentToStaticFromValueMethod(
+                final GeneratedSource generatedSource) {
+          Assumptions.assumeFalse(generatedSource.enumUnknownDefaultCase());
+          Assumptions.assumeTrue(String.class.equals(generatedSource.generatedFields()[0].type()));
+
+          CustomAssertions.assertStaticMethodThrowsWhenInvoked(
+              CustomAssertions.assertClassHasMethod(
+                  generatedSource.getClassUnderTest(),
+                  "fromValue",
+                  generatedSource.fieldClasses()[0]),
+              IllegalArgumentException.class,
+              "invalid");
+        }
+
+        @ParameterizedTest
+        @MethodSource(GENERATED_ENUM_TESTS_METHOD_SOURCE)
+        @DisplayName(
+            "Generated `static fromValue(T)` method ALWAYS `throw IllegalArgumentException` when non-matching integer `value` is given")
+        void
+            alwaysThrowIllegalArgumentExceptionWhenProvidingNonMatchingIntegerValueAsArgumentToStaticFromValueMethod(
+                final GeneratedSource generatedSource) {
+          Assumptions.assumeFalse(generatedSource.enumUnknownDefaultCase());
+          Assumptions.assumeTrue(Integer.class.equals(generatedSource.generatedFields()[0].type()));
+
+          CustomAssertions.assertStaticMethodThrowsWhenInvoked(
+              CustomAssertions.assertClassHasMethod(
+                  generatedSource.getClassUnderTest(),
+                  "fromValue",
+                  generatedSource.fieldClasses()[0]),
+              IllegalArgumentException.class,
+              42);
+        }
+
+        @ParameterizedTest
+        @MethodSource(GENERATED_ENUM_TESTS_METHOD_SOURCE)
+        @DisplayName(
+            "Generated `static fromValue(T)` method ALWAYS return expected `enum`-constant when `value` match")
+        void
+            alwaysReturnEnumConstantWhenProvidingExistingEnumValueAsArgumentToStaticFromValueMethod(
+                final GeneratedSource generatedSource) {
+          final Method fromValueMethod =
+              CustomAssertions.assertClassHasMethod(
+                  generatedSource.getClassUnderTest(),
+                  "fromValue",
+                  generatedSource.fieldClasses()[0]);
+          for (final GeneratedField<?> generatedField : generatedSource.generatedFields()) {
+            CustomAssertions.assertStaticMethodReturnsNonNull(
+                fromValueMethod, generatedField.enumValue());
+          }
+        }
+
+        @Nested
+        @DisplayName("Testing `<useEnumCaseInsensitive>false</useEnumCaseInsensitive>`")
+        class UseEnumCaseInsensitiveFalseTests {
+
+          @ParameterizedTest
+          @MethodSource(GENERATED_ENUM_TESTS_METHOD_SOURCE)
+          @DisplayName(
+              "Generated `static fromValue(T)` method throws `IllegalArgumentException` when string `value` has wrong case")
+          void
+              whenConfigOptionUseEnumCaseInsensitiveIsFalseThenFromValueMethodThrowsIllegalArgumentExceptionWhenGivenValueHasWrongCase(
+                  final GeneratedSource generatedSource) {
+            Assumptions.assumeFalse(generatedSource.enumUnknownDefaultCase());
+            Assumptions.assumeFalse(generatedSource.useEnumCaseInsensitive());
+            Assumptions.assumeTrue(
+                String.class.equals(generatedSource.generatedFields()[0].type()));
+
+            final Method fromValueMethod =
+                CustomAssertions.assertClassHasMethod(
+                    generatedSource.getClassUnderTest(),
+                    "fromValue",
+                    generatedSource.fieldClasses()[0]);
+            for (final GeneratedField<?> generatedField : generatedSource.generatedFields()) {
+              final Object enumValue = ((String) generatedField.enumValue()).toLowerCase();
+              CustomAssertions.assertStaticMethodThrowsWhenInvoked(
+                  fromValueMethod, IllegalArgumentException.class, enumValue);
+            }
+          }
+        }
+
+        @Nested
+        @DisplayName("Testing `<useEnumCaseInsensitive>true</useEnumCaseInsensitive>`")
+        class UseEnumCaseInsensitiveTrueTests {
+
+          @ParameterizedTest
+          @MethodSource(GENERATED_ENUM_TESTS_METHOD_SOURCE)
+          @DisplayName(
+              "Generated `static fromValue(T)` method returns `enum`-constant when string `value` has different case")
+          void
+              whenConfigOptionUseEnumCaseInsensitiveIsTrueThenFromValueMethodReturnsEnumConstantWithDifferentCase(
+                  final GeneratedSource generatedSource) {
+            Assumptions.assumeTrue(generatedSource.useEnumCaseInsensitive());
+            Assumptions.assumeTrue(
+                String.class.equals(generatedSource.generatedFields()[0].type()));
+
+            final Method fromValueMethod =
+                CustomAssertions.assertClassHasMethod(
+                    generatedSource.getClassUnderTest(),
+                    "fromValue",
+                    generatedSource.fieldClasses()[0]);
+            for (final GeneratedField<?> generatedField : generatedSource.generatedFields()) {
+              final Object enumValue = ((String) generatedField.enumValue()).toLowerCase();
+              CustomAssertions.assertStaticMethodReturnsNonNull(fromValueMethod, enumValue);
+            }
+          }
+        }
+      }
+
+      @Nested
+      @DisplayName("Testing `<enumUnknownDefaultCase>`")
+      class EnumUnknownDefaultCaseTests {
+
+        @Nested
+        @DisplayName("Testing `<enumUnknownDefaultCase>false</enumUnknownDefaultCase>`")
+        class EnumUnknownDefaultCaseFalseTests {
+
+          @ParameterizedTest
+          @MethodSource(GENERATED_ENUM_TESTS_METHOD_SOURCE)
+          @DisplayName("Generates an `enum` with expected number of constants")
+          void
+              whenConfigOptionEnumUnknownDefaultCaseIsFalseThenGeneratedEnumClassHasExpectedNumberOfConstants(
+                  final GeneratedSource generatedSource) {
+            Assumptions.assumeFalse(generatedSource.enumUnknownDefaultCase());
+            CustomAssertions.assertClassHasEnumConstants(
+                generatedSource.getClassUnderTest(), generatedSource.generatedFields().length);
+          }
+
+          @ParameterizedTest
+          @MethodSource(GENERATED_ENUM_TESTS_METHOD_SOURCE)
+          @DisplayName(
+              "Generated `enum`-class with `String` values does NOT have `\"UNKNOWN_DEFAULT_OPEN_API\"` as a constant")
+          void
+              whenConfigOptionEnumUnknownDefaultCaseIsFalseThenGeneratedEnumClassDoesNotHaveUnknownDefaultOpenApiEnumConstant(
+                  final GeneratedSource generatedSource) {
+            Assumptions.assumeFalse(generatedSource.enumUnknownDefaultCase());
+            Assumptions.assumeTrue(
+                String.class.equals(generatedSource.generatedFields()[0].type()));
+
+            CustomAssertions.assertClassDoesNotHaveEnumConstantWithName(
+                generatedSource.getClassUnderTest(), "UNKNOWN_DEFAULT_OPEN_API");
+          }
+
+          @ParameterizedTest
+          @MethodSource(GENERATED_ENUM_TESTS_METHOD_SOURCE)
+          @DisplayName(
+              "Generated `enum`-class with `Integer` values does NOT have `\"NUMBER_unknown_default_open_api\"` as a constant")
+          void
+              whenConfigOptionEnumUnknownDefaultCaseIsFalseThenGeneratedEnumClassDoesNotHaveNumberUnknownDefaultOpenApiEnumConstant(
+                  final GeneratedSource generatedSource) {
+            Assumptions.assumeFalse(generatedSource.enumUnknownDefaultCase());
+            Assumptions.assumeTrue(
+                Integer.class.equals(generatedSource.generatedFields()[0].type()));
+
+            CustomAssertions.assertClassDoesNotHaveEnumConstantWithName(
+                generatedSource.getClassUnderTest(), "NUMBER_unknown_default_open_api");
+          }
+
+          @ParameterizedTest
+          @MethodSource(GENERATED_ENUM_TESTS_METHOD_SOURCE)
+          @DisplayName(
+              "Generated `enum`-class with `URI` values does NOT have `\"UNKNOWN_DEFAULT_OPEN_API\"` as a constant")
+          void
+              whenConfigOptionEnumUnknownDefaultCaseIsFalseThenGeneratedEnumClassDoesNotHaveUriUnknownDefaultOpenApiEnumConstant(
+                  final GeneratedSource generatedSource) {
+            Assumptions.assumeFalse(generatedSource.enumUnknownDefaultCase());
+            Assumptions.assumeTrue(URI.class.equals(generatedSource.generatedFields()[0].type()));
+
+            CustomAssertions.assertClassDoesNotHaveEnumConstantWithName(
+                generatedSource.getClassUnderTest(), "UNKNOWN_DEFAULT_OPEN_API");
+          }
+
+          @ParameterizedTest
+          @MethodSource(GENERATED_ENUM_TESTS_METHOD_SOURCE)
+          @DisplayName(
+              "Generated `static fromValue(T)` method ALWAYS `throw IllegalArgumentException` when `null` is given")
+          void
+              alwaysThrowIllegalArgumentExceptionWhenProvidingNullAsArgumentToStaticFromValueMethod(
+                  final GeneratedSource generatedSource) {
+            Assumptions.assumeFalse(generatedSource.enumUnknownDefaultCase());
+
+            CustomAssertions.assertStaticMethodThrowsWhenInvoked(
+                CustomAssertions.assertClassHasMethod(
+                    generatedSource.getClassUnderTest(),
+                    "fromValue",
+                    generatedSource.fieldClasses()[0]),
+                IllegalArgumentException.class,
+                (Object) null);
+          }
+        }
+
+        @Nested
+        @DisplayName("Testing `<enumUnknownDefaultCase>true</enumUnknownDefaultCase>`")
+        class EnumUnknownDefaultCaseTrueTests {
+
+          @ParameterizedTest
+          @MethodSource(GENERATED_ENUM_TESTS_METHOD_SOURCE)
+          @DisplayName(
+              "Generates an `enum` with expected number of constants PLUS an additional one constant")
+          void
+              whenConfigOptionEnumUnknownDefaultCaseIsTrueThenGeneratedEnumClassHasExpectedNumberOfConstantsPlusOne(
+                  final GeneratedSource generatedSource) {
+            Assumptions.assumeTrue(generatedSource.enumUnknownDefaultCase());
+            CustomAssertions.assertClassHasEnumConstants(
+                generatedSource.getClassUnderTest(), generatedSource.generatedFields().length + 1);
+          }
+
+          @ParameterizedTest
+          @MethodSource(GENERATED_ENUM_TESTS_METHOD_SOURCE)
+          @DisplayName(
+              "Generated `enum`-class with `String` values HAS `\"UNKNOWN_DEFAULT_OPEN_API\"` as a constant")
+          void
+              whenConfigOptionEnumUnknownDefaultCaseIsTrueThenGeneratedEnumClassHasUnknownDefaultOpenApiEnumConstant(
+                  final GeneratedSource generatedSource) {
+            Assumptions.assumeTrue(generatedSource.enumUnknownDefaultCase());
+            Assumptions.assumeTrue(
+                String.class.equals(generatedSource.generatedFields()[0].type()));
+
+            CustomAssertions.assertClassHasEnumConstantWithName(
+                generatedSource.getClassUnderTest(), "UNKNOWN_DEFAULT_OPEN_API");
+          }
+
+          @ParameterizedTest
+          @MethodSource(GENERATED_ENUM_TESTS_METHOD_SOURCE)
+          @DisplayName(
+              "Generated `enum`-class with `String` values HAS `\"UNKNOWN_DEFAULT_OPEN_API\"` constant with `value` as `\"unknown_default_open_api\"`")
+          void
+              whenConfigOptionEnumUnknownDefaultCaseIsTrueThenGeneratedEnumClassHasUnknownDefaultOpenApiEnumConstantWithExpectedStringValue(
+                  final GeneratedSource generatedSource) {
+            Assumptions.assumeTrue(generatedSource.enumUnknownDefaultCase());
+            Assumptions.assumeTrue(
+                String.class.equals(generatedSource.generatedFields()[0].type()));
+
+            CustomAssertions.assertClassHasEnumConstantWithValue(
+                generatedSource.getClassUnderTest(), "unknown_default_open_api");
+          }
+
+          @ParameterizedTest
+          @MethodSource(GENERATED_ENUM_TESTS_METHOD_SOURCE)
+          @DisplayName(
+              "Generated `enum`-class with `URI` values HAS `\"UNKNOWN_DEFAULT_OPEN_API\"` as a constant")
+          void
+              whenConfigOptionEnumUnknownDefaultCaseIsTrueThenGeneratedEnumClassHasUriUnknownDefaultOpenApiEnumConstant(
+                  final GeneratedSource generatedSource) {
+            Assumptions.assumeTrue(generatedSource.enumUnknownDefaultCase());
+            Assumptions.assumeTrue(URI.class.equals(generatedSource.generatedFields()[0].type()));
+
+            CustomAssertions.assertClassHasEnumConstantWithName(
+                generatedSource.getClassUnderTest(), "UNKNOWN_DEFAULT_OPEN_API");
+          }
+
+          @ParameterizedTest
+          @MethodSource(GENERATED_ENUM_TESTS_METHOD_SOURCE)
+          @DisplayName(
+              "Generated `enum`-class with `URI` values HAS `\"UNKNOWN_DEFAULT_OPEN_API\"` constant with `value` as `URI.create(\"11184809\")`")
+          void
+              whenConfigOptionEnumUnknownDefaultCaseIsTrueThenGeneratedEnumClassHasUnknownDefaultOpenApiEnumConstantWithExpectedUriValue(
+                  final GeneratedSource generatedSource) {
+            Assumptions.assumeTrue(generatedSource.enumUnknownDefaultCase());
+            Assumptions.assumeTrue(URI.class.equals(generatedSource.generatedFields()[0].type()));
+
+            CustomAssertions.assertClassHasEnumConstantWithValue(
+                generatedSource.getClassUnderTest(), URI.create("11184809"));
+          }
+
+          @ParameterizedTest
+          @MethodSource(GENERATED_ENUM_TESTS_METHOD_SOURCE)
+          @DisplayName(
+              "Generated `enum`-class with `Integer` values HAS `\"NUMBER_unknown_default_open_api\"` as a constant")
+          void
+              whenConfigOptionEnumUnknownDefaultCaseIsTrueThenGeneratedEnumClassHasNumberUnknownDefaultOpenApiEnumConstant(
+                  final GeneratedSource generatedSource) {
+            Assumptions.assumeTrue(generatedSource.enumUnknownDefaultCase());
+            Assumptions.assumeTrue(
+                Integer.class.equals(generatedSource.generatedFields()[0].type()));
+
+            CustomAssertions.assertClassHasEnumConstantWithName(
+                generatedSource.getClassUnderTest(), "NUMBER_unknown_default_open_api");
+          }
+
+          @ParameterizedTest
+          @MethodSource(GENERATED_ENUM_TESTS_METHOD_SOURCE)
+          @DisplayName(
+              "Generated `enum`-class with `Integer` values HAS `\"NUMBER_unknown_default_open_api\"` constant with `value` as `11184809`")
+          void
+              whenConfigOptionEnumUnknownDefaultCaseIsTrueThenGeneratedEnumClassHasUnknownDefaultOpenApiEnumConstantWithExpectedIntegerValue(
+                  final GeneratedSource generatedSource) {
+            Assumptions.assumeTrue(generatedSource.enumUnknownDefaultCase());
+            Assumptions.assumeTrue(
+                Integer.class.equals(generatedSource.generatedFields()[0].type()));
+
+            CustomAssertions.assertClassHasEnumConstantWithValue(
+                generatedSource.getClassUnderTest(), 11184809);
+          }
+
+          @ParameterizedTest
+          @MethodSource(GENERATED_ENUM_TESTS_METHOD_SOURCE)
+          @DisplayName(
+              "Generated `static fromValue(T)` method returns `NUMBER_unknown_default_open_api` when `null` is given")
+          void whenConfigOptionEnumUnknownDefaultCaseIsTrueThenFromValueReturnsIntegerDefaultValue(
+              final GeneratedSource generatedSource) {
+            Assumptions.assumeTrue(generatedSource.enumUnknownDefaultCase());
+            Assumptions.assumeTrue(
+                Integer.class.equals(generatedSource.generatedFields()[0].type()));
+            CustomAssertions.assertStaticMethodReturnsNonNull(
+                CustomAssertions.assertClassHasMethod(
+                    generatedSource.getClassUnderTest(),
+                    "fromValue",
+                    generatedSource.fieldClasses()[0]),
+                (Object) null);
+          }
+
+          @ParameterizedTest
+          @MethodSource(GENERATED_ENUM_TESTS_METHOD_SOURCE)
+          @DisplayName(
+              "Generated `static fromValue(T)` method returns `\"NUMBER_unknown_default_open_api\"` when `null` is given")
+          void whenConfigOptionEnumUnknownDefaultCaseIsTrueThenFromValueReturnsStringDefaultValue(
+              final GeneratedSource generatedSource) {
+            Assumptions.assumeTrue(generatedSource.enumUnknownDefaultCase());
+            Assumptions.assumeTrue(
+                String.class.equals(generatedSource.generatedFields()[0].type()));
+
+            CustomAssertions.assertStaticMethodReturnsNonNull(
+                CustomAssertions.assertClassHasMethod(
+                    generatedSource.getClassUnderTest(),
+                    "fromValue",
+                    generatedSource.fieldClasses()[0]),
+                (Object) null);
+          }
+
+          @ParameterizedTest
+          @MethodSource(GENERATED_ENUM_TESTS_METHOD_SOURCE)
+          @DisplayName(
+              "Generated `static fromValue(T)` method returns `\"UNKNOWN_DEFAULT_OPEN_API\"` when `null` is given")
+          void whenConfigOptionEnumUnknownDefaultCaseIsTrueThenFromValueReturnsUriDefaultValue(
+              final GeneratedSource generatedSource) {
+            Assumptions.assumeTrue(generatedSource.enumUnknownDefaultCase());
+            Assumptions.assumeTrue(URI.class.equals(generatedSource.generatedFields()[0].type()));
+
+            CustomAssertions.assertStaticMethodReturnsNonNull(
+                CustomAssertions.assertClassHasMethod(
+                    generatedSource.getClassUnderTest(),
+                    "fromValue",
+                    generatedSource.fieldClasses()[0]),
+                (Object) null);
+          }
         }
       }
     }
